Modified ServerTabWidget; Added Utility namespace
Changed ServerTabWidget to include a QTextBrowser for world information,
and a QListWidget for connected players. These additions aren't yet
functional.

Created the namespace "Utility" for commonly used functions.
Functions added to Utility:
left - Returns everything left of a separator in a QString
right - Returns everything right of a separator in a QString
getDirectoryEntryList - Returns a QStringList with file names in a
directory.